requirements.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
django-filter==2.2.0
dynamic-rest==1.9.6
mysqlclient==1.4.4
ujson==1.35
requests==2.22.0
django-storages[google]==1.8
gunicorn==20.0.4
django-widget-tweaks==1.4.5
sorl-thumbnail==12.6.2
Pillow==7.0.0
whitenoise==5.1.0


# Jupyter Notebooks
# env DJANGO_ALLOW_ASYNC_UNSAFE=true ./manage.py shell_plus --notebook
# jupyter
# ipython
# django-extensions
